+++ src/types/iri.rs
@@ -1,6 +1,9 @@
 //! IRI-specific implementations.
 
 #[cfg(feature = "alloc")]
+use alloc::string::String;
+
+#[cfg(feature = "alloc")]
 use crate::convert::{try_percent_encode_iri_inline, MappedToUri};
 use crate::spec::IriSpec;
 #[cfg(feature = "alloc")]
@@ -10,6 +13,7 @@ use crate::types::{RiAbsoluteStr, RiFragmentStr, RiReferenceStr, RiRelativeStr, 
 use crate::types::{
     RiAbsoluteString, RiFragmentString, RiReferenceString, RiRelativeString, RiString,
 };
+use crate::types::{UriAbsoluteStr, UriFragmentStr, UriReferenceStr, UriRelativeStr, UriStr};
 #[cfg(feature = "alloc")]
 use crate::types::{
     UriAbsoluteString, UriFragmentString, UriReferenceString, UriRelativeString, UriString,
@@ -56,7 +60,7 @@ pub type IriRelativeStr = RiRelativeStr<IriSpec>;
 pub type IriRelativeString = RiRelativeString<IriSpec>;
 
 /// Implements the conversion from an IRI into a URI.
-macro_rules! impl_encode_to_uri {
+macro_rules! impl_conversion_between_uri {
     (
         $ty_owned_iri:ident,
         $ty_owned_uri:ident,
@@ -66,7 +70,6 @@ macro_rules! impl_encode_to_uri {
         $example_uri:expr
     ) => {
         /// Conversion from an IRI into a URI.
-        #[cfg(feature = "alloc")]
         impl $ty_borrowed_iri {
             /// Percent-encodes the IRI into a valid URI that identifies the equivalent resource.
             ///
@@ -77,7 +80,7 @@ macro_rules! impl_encode_to_uri {
             ///
             /// ```
             /// # use iri_string::validate::Error;
-            /// #[cfg(feature = "alloc")] {
+            /// # #[cfg(feature = "alloc")] {
             #[doc = concat!("use iri_string::types::{", stringify!($ty_borrowed_iri), ", ", stringify!($ty_owned_uri), "};")]
             ///
             #[doc = concat!("let iri = ", stringify!($ty_borrowed_iri), "::new(", stringify!($example_iri), ")?;")]
@@ -87,13 +90,54 @@ macro_rules! impl_encode_to_uri {
             /// # }
             /// # Ok::<_, Error>(())
             /// ```
+            #[cfg(feature = "alloc")]
+            #[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
             pub fn encode_to_uri(&self) -> $ty_owned_uri {
                 MappedToUri::from(self).allocate_and_write().expect("failed to allocate memory")
+            }
+
+            /// Converts an IRI into a URI without modification, if possible.
+            ///
+            /// This is semantically equivalent to
+            #[doc = concat!("`", stringify!($ty_borrowed_uri), "::new(self.as_str()).ok()`.")]
+            ///
+            /// # Examples
+            ///
+            /// ```
+            /// # use iri_string::validate::Error;
+            #[doc = concat!("use iri_string::types::{", stringify!($ty_borrowed_iri), ", ", stringify!($ty_borrowed_uri), "};")]
+            ///
+            #[doc = concat!("let ascii_iri = ", stringify!($ty_borrowed_iri), "::new(", stringify!($example_uri), ")?;")]
+            /// assert_eq!(
+            ///     ascii_iri.as_uri().map(AsRef::as_ref),
+            #[doc = concat!("    Some(", stringify!($example_uri), ")")]
+            /// );
+            ///
+            #[doc = concat!("let nonascii_iri = ", stringify!($ty_borrowed_iri), "::new(", stringify!($example_iri), ")?;")]
+            /// assert_eq!(nonascii_iri.as_uri(), None);
+            /// # Ok::<_, Error>(())
+            /// ```
+            pub fn as_uri(&self) -> Option<&$ty_borrowed_uri> {
+                if !self.as_str().is_ascii() {
+                    return None;
+                }
+                debug_assert!(
+                    <$ty_borrowed_uri>::new(self.as_str()).is_ok(),
+                    "[consistency] the ASCII-only IRI must also be a valid URI"
+                );
+                let uri = unsafe {
+                    // SAFETY: An ASCII-only IRI is a URI.
+                    // URI (by `UriSpec`) is a subset of IRI (by `IriSpec`),
+                    // and the difference is that URIs can only have ASCII characters.
+                    <$ty_borrowed_uri>::new_maybe_unchecked(self.as_str())
+                };
+                Some(uri)
             }
         }
 
         /// Conversion from an IRI into a URI.
         #[cfg(feature = "alloc")]
+        #[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
         impl $ty_owned_iri {
             /// Percent-encodes the IRI into a valid URI that identifies the equivalent resource.
             ///
@@ -152,11 +196,50 @@ macro_rules! impl_encode_to_uri {
             pub fn encode_into_uri(self) -> $ty_owned_uri {
                 MappedToUri::from(self.as_slice()).allocate_and_write().expect("failed to allocate memory")
             }
+
+            /// Converts an IRI into a URI without modification, if possible.
+            ///
+            /// # Examples
+            ///
+            /// ```
+            /// # use iri_string::validate::Error;
+            #[doc = concat!("use iri_string::types::{", stringify!($ty_owned_iri), ", ", stringify!($ty_owned_uri), "};")]
+            ///
+            #[doc = concat!("let ascii_iri = ", stringify!($ty_owned_iri), "::try_from(", stringify!($example_uri), ")?;")]
+            /// assert_eq!(
+            ///     ascii_iri.try_into_uri().map(|uri| uri.to_string()),
+            #[doc = concat!("    Ok(", stringify!($example_uri), ".to_string())")]
+            /// );
+            ///
+            #[doc = concat!("let nonascii_iri = ", stringify!($ty_owned_iri), "::try_from(", stringify!($example_iri), ")?;")]
+            /// assert_eq!(
+            ///     nonascii_iri.try_into_uri().map_err(|iri| iri.to_string()),
+            #[doc = concat!("    Err(", stringify!($example_iri), ".to_string())")]
+            /// );
+            /// # Ok::<_, Error>(())
+            /// ```
+            pub fn try_into_uri(self) -> Result<$ty_owned_uri, $ty_owned_iri> {
+                if !self.as_str().is_ascii() {
+                    return Err(self);
+                }
+                let s: String = self.into();
+                debug_assert!(
+                    <$ty_borrowed_uri>::new(s.as_str()).is_ok(),
+                    "[consistency] the ASCII-only IRI must also be a valid URI"
+                );
+                let uri = unsafe {
+                    // SAFETY: An ASCII-only IRI is a URI.
+                    // URI (by `UriSpec`) is a subset of IRI (by `IriSpec`),
+                    // and the difference is that URIs can only have ASCII characters.
+                    <$ty_owned_uri>::new_maybe_unchecked(s)
+                };
+                Ok(uri)
+            }
         }
     };
 }
 
-impl_encode_to_uri!(
+impl_conversion_between_uri!(
     IriAbsoluteString,
     UriAbsoluteString,
     IriAbsoluteStr,
@@ -164,7 +247,7 @@ impl_encode_to_uri!(
     "http://example.com/?alpha=\u{03B1}",
     "http://example.com/?alpha=%CE%B1"
 );
-impl_encode_to_uri!(
+impl_conversion_between_uri!(
     IriReferenceString,
     UriReferenceString,
     IriReferenceStr,
@@ -172,7 +255,7 @@ impl_encode_to_uri!(
     "http://example.com/?alpha=\u{03B1}",
     "http://example.com/?alpha=%CE%B1"
 );
-impl_encode_to_uri!(
+impl_conversion_between_uri!(
     IriRelativeString,
     UriRelativeString,
     IriRelativeStr,
@@ -180,7 +263,7 @@ impl_encode_to_uri!(
     "../?alpha=\u{03B1}",
     "../?alpha=%CE%B1"
 );
-impl_encode_to_uri!(
+impl_conversion_between_uri!(
     IriString,
     UriString,
     IriStr,
@@ -188,7 +271,7 @@ impl_encode_to_uri!(
     "http://example.com/?alpha=\u{03B1}",
     "http://example.com/?alpha=%CE%B1"
 );
-impl_encode_to_uri!(
+impl_conversion_between_uri!(
     IriFragmentString,
     UriFragmentString,
     IriFragmentStr,
@@ -196,3 +279,32 @@ impl_encode_to_uri!(
     "alpha-is-\u{03B1}",
     "alpha-is-%CE%B1"
 );
+
+#[cfg(test)]
+#[cfg(feature = "alloc")]
+mod tests {
+    use crate::types::{IriReferenceString, UriReferenceStr};
+
+    const CASES: &[(&str, &str)] = &[
+        ("?alpha=\u{03B1}", "?alpha=%CE%B1"),
+        (
+            "?katakana-letter-i=\u{30A4}",
+            "?katakana-letter-i=%E3%82%A4",
+        ),
+        ("?sushi=\u{1f363}", "?sushi=%F0%9F%8D%A3"),
+    ];
+
+    #[test]
+    fn iri_to_uri() {
+        for (iri, expected_uri) in CASES {
+            let expected_uri =
+                UriReferenceStr::new(*expected_uri).expect("test cases must be valid");
+
+            let mut iri = IriReferenceString::try_from(*iri).expect("test cases must be valid");
+            iri.encode_to_uri();
+            assert_eq!(iri, expected_uri);
+            iri.encode_to_uri();
+            assert_eq!(iri, expected_uri, "`encode_to_uri` must be idempotent");
+        }
+    }
+}